style: minor code compartmentation
