+++ requirements-dev.txt
@@ -1,6 +1,9 @@
 black==22.10.0
 coverage-badge==1.1.0
 flake8==5.0.4
+flake8-print>=5.0.0
+flake8-import-order>=0.18.0
+flake8-comprehensions>=3.10.0
 gitlint==0.17.0
 mock==4.0.3
 mypy==0.990
